Actualizar idioma a español en index.html; modificar lógica de habilitación de botón en HistorialMantenimiento; simplificar TablaClientes y ajustar encabezados en TablaMantenimiento según rol de usuario.
